@@ -12,9 +12,17 @@ import SwiftAudioPlayer
 struct AudioInfo: Hashable {
     var index: Int = 0
     
-    var urls: [URL] = [URL(string: "https://www.fesliyanstudios.com/musicfiles/2019-04-23_-_Trusted_Advertising_-_www.fesliyanstudios.com/15SecVersion2019-04-23_-_Trusted_Advertising_-_www.fesliyanstudios.com.mp3")!,
-                       URL(string: "https://chtbl.com/track/18338/traffic.libsyn.com/secure/acquired/acquired_-_armrev_2.mp3?dest-id=376122")!,
-                       URL(string: "https://ice6.somafm.com/groovesalad-256-mp3")!]
+  var urls: [URL] = [
+    URL(
+      string:
+        "https://www.fesliyanstudios.com/musicfiles/2019-04-23_-_Trusted_Advertising_-_www.fesliyanstudios.com/15SecVersion2019-04-23_-_Trusted_Advertising_-_www.fesliyanstudios.com.mp3"
+    )!,
+    URL(
+      string:
+        "https://chtbl.com/track/18338/traffic.libsyn.com/secure/acquired/acquired_-_armrev_2.mp3?dest-id=376122"
+    )!,
+    URL(string: "https://ice6.somafm.com/groovesalad-256-mp3")!,
+  ]
     
     var url: URL {
         switch index {
@@ -43,18 +51,16 @@ struct AudioInfo: Hashable {
     }
     
     let artist: String = "SwiftAudioPlayer Sample App"
-    let releaseDate: Int = 1550790640
+  let releaseDate: Int = 1_550_790_640
     
     var lockscreenInfo: SALockScreenInfo {
-        get {
-            return SALockScreenInfo(title: self.title, artist: self.artist, albumTitle: nil, artwork: nil, releaseDate: self.releaseDate)
-        }
+    return SALockScreenInfo(
+      title: self.title, artist: self.artist, albumTitle: nil, artwork: nil,
+      releaseDate: self.releaseDate)
     }
     
     var savedUrl: URL? {
-        get {
             return savedUrls[index]
-        }
     }
     
     var savedUrls: [URL?] = [nil, nil, nil]
