@@ -1,5 +1,5 @@
-import XCTest
 import SwiftAudioPlayer
+import XCTest
 
 class Tests: XCTestCase {
     
@@ -20,7 +20,7 @@ class Tests: XCTestCase {
     
     func testPerformanceExample() {
         // This is an example of a performance test case.
-        self.measure() {
+    self.measure {
             // Put the code you want to measure the time of here.
         }
     }
