@@ -29,11 +29,14 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
-import Foundation
 import AVFoundation
+import Foundation
 
-
-func ParserPropertyListener(_ context: UnsafeMutableRawPointer, _ streamId: AudioFileStreamID, _ propertyId: AudioFileStreamPropertyID, _ flags: UnsafeMutablePointer<AudioFileStreamPropertyFlags>) {
+func ParserPropertyListener(
+  _ context: UnsafeMutableRawPointer, _ streamId: AudioFileStreamID,
+  _ propertyId: AudioFileStreamPropertyID,
+  _ flags: UnsafeMutablePointer<AudioFileStreamPropertyFlags>
+) {
     let selfAudioParser = Unmanaged<AudioParser>.fromOpaque(context).takeUnretainedValue()
     
     Log.info("audio file stream property: \(propertyId.description)")
@@ -48,13 +51,15 @@ func ParserPropertyListener(_ context: UnsafeMutableRawPointer, _ streamId: Audi
         break
     case kAudioFileStreamProperty_AudioDataByteCount:
         GetPropertyValue(&selfAudioParser.parsedAudioPacketDataSize, streamId, propertyId)
-        selfAudioParser.expectedFileSizeInBytes = selfAudioParser.parsedAudioDataOffset + selfAudioParser.parsedAudioPacketDataSize
-        break;
+    selfAudioParser.expectedFileSizeInBytes =
+      selfAudioParser.parsedAudioDataOffset + selfAudioParser.parsedAudioPacketDataSize
+    break
     case kAudioFileStreamProperty_DataOffset:
         GetPropertyValue(&selfAudioParser.parsedAudioDataOffset, streamId, propertyId)
         
-        if(selfAudioParser.parsedAudioPacketDataSize != 0) {
-            selfAudioParser.expectedFileSizeInBytes = selfAudioParser.parsedAudioDataOffset + selfAudioParser.parsedAudioPacketDataSize
+    if selfAudioParser.parsedAudioPacketDataSize != 0 {
+      selfAudioParser.expectedFileSizeInBytes =
+        selfAudioParser.parsedAudioDataOffset + selfAudioParser.parsedAudioPacketDataSize
         }
         
         break
@@ -64,7 +69,9 @@ func ParserPropertyListener(_ context: UnsafeMutableRawPointer, _ streamId: Audi
 }
 
 //property is like the medatada of
-func GetPropertyValue<T>(_ value: inout T, _ streamId: AudioFileStreamID, _ propertyId: AudioFileStreamPropertyID) {
+func GetPropertyValue<T>(
+  _ value: inout T, _ streamId: AudioFileStreamID, _ propertyId: AudioFileStreamPropertyID
+) {
     var propertySize: UInt32 = 0
     guard AudioFileStreamGetPropertyInfo(streamId, propertyId, &propertySize, nil) == noErr else {//try to get the size of the property
         Log.monitor("failed to get info for property:\(propertyId.description)")
