@@ -25,20 +25,16 @@
 
 import Foundation
 
-
-/**
- P for payload
- */
+/// P for payload
 class DirectorThreadSafeClosures<P>  {
     typealias TypeClosure = (P) throws -> Void
-    private var queue: DispatchQueue = DispatchQueue(label: "SwiftAudioPlayer.thread_safe_map", attributes: .concurrent)
+  private var queue: DispatchQueue = DispatchQueue(
+    label: "SwiftAudioPlayer.thread_safe_map", attributes: .concurrent)
     private var closures: [UInt: TypeClosure] = [:]
     private var cache: P? = nil
     
     var count: Int {
-        get {
             return closures.count
-        }
     }
     
     func resetCache() {
