@@ -9,22 +9,17 @@ import Foundation
 
 extension URL {
     var key: String {
-        get {
             return "audio_\(self.absoluteString.hashed)"
         }
     }
-}
 
-
-fileprivate extension String {
-    var hashed: UInt64 {
-        get {
+extension String {
+  fileprivate var hashed: UInt64 {
             var result = UInt64 (8742)
             let buf = [UInt8](self.utf8)
             for b in buf {
-                result = 127 * (result & 0x00ffffffffffffff) + UInt64(b)
+      result = 127 * (result & 0x00ff_ffff_ffff_ffff) + UInt64(b)
             }
             return result
         }
     }
-}
